Remove attempt files and rename crate
