+++ crates/builder/Builder.toml
@@ -1,16 +1,43 @@
-[postbuild.assembly.target.profile.plugin.action1]
+[postbuild.web.aarch64-apple-darwin.custom.plugin.action1]
 test = "lorem"
 test2 = "ipsum"
 test3 = "dolor"
 
-[postbuild.assembly.target.profile.plugin.action2]
+[postbuild.web.aarch64-apple-darwin.custom.plugin.action2]
 test = "lorem"
 test2 = "ipsum"
 
-[postbuild.assembly.target.profile.plugin.overrided]
+[postbuild.web.aarch64-apple-darwin.custom2.plugin.action2]
 test = "lorem"
 test2 = "ipsum"
+
+[postbuild.web.aarch64-apple-darwin.custom.plugin.action3]
+test = "lorem"
+test2 = "ipsum"
+
+[postbuild.web.i686-pc-windows-msvc.custom.plugin.action3]
+test = "lorem"
+test2 = "ipsum"
+
+[install.i686-pc-windows-msvc.plugin]
+cmd-alias = "plugin"
+install = "cargo_cmd_package_name"
 
 [install.clippy]
+cmd-alias = "cargo-clippy"
 version = "1"
-install = "clippy_url"
+install = "cargo_cmd_package_name"
+
+[install.aarch64-apple-darwin.plugin]
+cmd-alias = "plugin"
+install = "cargo_cmd_package_name"
+
+[install.i686-pc-windows-msvc.plugin2]
+cmd-alias = "plugin"
+install = "cargo_cmd_package_name"
+
+[githook.pre-commit]
+plugin = {}
+
+[githook.post-commit]
+plugin = {}
