+++ crates/builder/Cargo.toml
@@ -3,6 +3,12 @@ edition = "2021"
 name = "builder"
 version = "0.1.0"
 
+[profile.custom]
+opt-level = 1
+
+[profile.custom2]
+opt-level = 1
+
 [dependencies]
 anyhow = "1.0"
 cargo_metadata = "0.18"
@@ -10,10 +16,9 @@ clap = { version = "4.5", features = ["derive", "env"] }
 serde = "1.0"
 serde_json = "1.0"
 serde_yaml = "0.9"
+target-lexicon = "0.12.15"
 toml = "0.8"
+walkdir = "2.5.0"
 which = "6.0.1"
 
-
-[package.metadata.postbuild.assembly.target.profile.plugin.overrided]
-test = "lorem"
-test2 = "ipsum"
+builder-poc = { workspace = true }
