--- crates/builder/src/output.rs
@@ -1,6 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Default, Serialize, Deserialize)]
-pub struct Output {
-    artifacts: Vec<String>,
-}
