+++ crates/builder/src/types/plugin/assembly.rs
