+++ crates/builder/src/types/plugin/installer/binary.rs
