+++ crates/builder/src/types/plugin/installer/mod.rs
